Move config types into snage package
