+++ packages/snage/src/query/match.ts
@@ -1,9 +1,10 @@
-import {Field, Note} from '../../../shared/type';
+import {Note} from '../../../shared/type';
 import {Expression, Operator, SingleExpression} from './parser';
 import {expectNever, ffVersionRegex} from '../util/util';
 import semver from 'semver';
 import stringsimi from 'string-similarity';
 import {zip} from 'fp-ts/lib/Array';
+import {Field} from '../config/type';
 
 interface Matcher {
     (n: Note): boolean;
